test: extract shared helpers to tests/helpers.mjs (#1440)
